# docs/reference/openapi.json
{
  "components": {
    "schemas": {
...
      "CountOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count"
        ],
        "title": "CountOut",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "The JSON body of every error response.",
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "issues": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issues"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "results": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Results"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "Event": {
        "description": "A stored event: NewEvent's fields (effective_at required) plus the ledger-assigned ones.",
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "causation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Causation Id"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "effective_at": {
            "format": "date-time",
            "title": "Effective At",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "prev_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Hash"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "stream_id": {
            "title": "Stream Id",
            "type": "string"
          },
          "stream_type": {
            "title": "Stream Type",
            "type": "string"
          },
          "stream_version": {
            "title": "Stream Version",
            "type": "integer"
          }
        },
        "required": [
          "event_type",
          "payload",
          "seq",
          "event_id",
          "stream_id",
          "stream_type",
          "stream_version",
          "scope",
          "actor",
          "recorded_at",
          "effective_at",
          "correlation_id",
          "causation_id",
          "source",
          "prev_hash",
          "hash"
        ],
        "title": "Event",
        "type": "object"
      },
...
      "RecordOut": {
        "description": "A record envelope (brief 6.2): the dict ``tl_core.services.queries`` returns.",
        "properties": {
          "conformance": {
            "title": "Conformance",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "effective_schema_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Schema Hash"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "last_seq": {
            "title": "Last Seq",
            "type": "integer"
          },
          "psets": {
            "additionalProperties": true,
            "title": "Psets",
            "type": "object"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "voided": {
            "title": "Voided",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "key",
          "type",
          "scope",
          "title",
          "description",
          "status",
          "psets",
          "voided",
          "version",
          "last_seq",
          "effective_schema_hash",
          "conformance",
          "created_at",
          "updated_at"
        ],
        "title": "RecordOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Static dev token (ADR-0005)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Records, links, workflow, files and the change feed over HTTP. Phase 0 is a development server: identity is a static bearer token and every authenticated actor may do everything (ADR-0005). Errors are `{error, message, ...}` bodies; the `error` codes are stable.",
    "title": "Throughline API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/records": {
      "get": {
        "description": "Records of a scope, filtered by the query language, ordered and paged.",
        "operationId": "list_records",
        "parameters": [
          {
            "description": "`company` or `project:<id>`.",
            "in": "query",
            "name": "scope",
            "required": true,
            "schema": {
              "description": "`company` or `project:<id>`.",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "description": "Filter in the query language (`status:open linked:NCR`).",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter in the query language (`status:open linked:NCR`).",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "record_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Record Type"
            }
          },
          {
            "description": "Same as `status:<value>` in `q`.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Same as `status:<value>` in `q`.",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "include_voided",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Voided",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Comma-separated `column[:asc|:desc]`; psets as `psets.<pset>.<prop>`.",
            "in": "query",
            "name": "order_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated `column[:asc|:desc]`; psets as `psets.<pset>.<prop>`.",
              "title": "Order By"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecordOut"
                  },
                  "title": "Response List Records",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Records",
        "tags": [
          "records"
        ]
      }
    },
    "/records/count": {
      "get": {
        "description": "How many records match (`limit`, `offset` and ordering do not apply).",
        "operationId": "count_records",
        "parameters": [
          {
            "description": "`company` or `project:<id>`.",
            "in": "query",
            "name": "scope",
            "required": true,
            "schema": {
              "description": "`company` or `project:<id>`.",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "description": "Filter in the query language (`status:open linked:NCR`).",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter in the query language (`status:open linked:NCR`).",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "record_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Record Type"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "include_voided",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Voided",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Count Records",
        "tags": [
          "records"
        ]
      }
    },
    "/records/lookup": {
      "get": {
        "description": "The record with this key in this scope (voided ones included). 404 when there is none.",
        "operationId": "lookup_record",
        "parameters": [
          {
            "description": "`company` or `project:<id>`.",
            "in": "query",
            "name": "scope",
            "required": true,
            "schema": {
              "description": "`company` or `project:<id>`.",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "description": "The record's key, e.g. `P123-NCR-0042`.",
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "description": "The record's key, e.g. `P123-NCR-0042`.",
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Lookup Record",
        "tags": [
          "records"
        ]
      }
    },
    "/records/{record_id}": {
      "get": {
        "description": "One record by id. The `ETag` header is the stream version (brief 11.1).",
        "operationId": "get_record",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Record",
        "tags": [
          "records"
        ]
      }
    },
...
    "/records/{record_id}/history": {
      "get": {
        "description": "Every event of the record's stream, oldest first. 404 when the record is unknown.",
        "operationId": "get_record_history",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  },
                  "title": "Response Get Record History",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Record History",
        "tags": [
          "records"
        ]
      }
    },
...
  }
}